+++ src/file/CSVFile.h
@@ -11,7 +11,7 @@
 class CSVFile : public File
 {
 private:
-	size_t row_count;
+	int row_count;
 
 public:
 	explicit CSVFile(const std::filesystem::path& file_path);
@@ -20,8 +20,8 @@ public:
 
 	CSVFile(std::filesystem::path new_path,
 			const Timestamp& new_last_write_time,
-			size_t new_size,
-			size_t new_row_count);
+			int new_size,
+			int new_row_count);
 
 	std::ostream& print(std::ostream& os) const override;
 
@@ -30,4 +30,4 @@ public:
 	bool MatchRowCount(const IntTerm& term) const override;
 };
 
-size_t get_row_count(const std::filesystem::path& path);
+int get_row_count(const std::filesystem::path& path);
